🚚 move tests/nightly/core to tests/nightly/utils
