+++ tests/nightly/utils/callbacks/compress_callback/dummy_config.yml
