+++ tests/nightly/utils/callbacks/compress_callback/dummy_lightning_model.py
