+++ tests/nightly/utils/callbacks/visualizer_callback/dummy_lightning_model.py
