+++ tests/nightly/utils/callbacks/visualizer_callback/test_visualizer.py
